% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{rgoenorm}
\alias{rgoenorm}
\title{Sample from the multivariate normal distribution using a covariance matrix Sigma drawn
from the Gaussian Orthogonal Ensemble, using m random variables. The mean_vector shall be 0}
\usage{
rgoenorm(n, m)
}
\description{
Sample from the multivariate normal distribution using a covariance matrix Sigma drawn
from the Gaussian Orthogonal Ensemble, using m random variables. The mean_vector shall be 0
}
