% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{rmvnorm}
\alias{rmvnorm}
\title{Random variable from the multivariate normal distribution
Sigma is the covariance matrix
n is the number of observations to sample
mean_vector is the mean vector}
\usage{
rmvnorm(n, sigma, mean_vector)
}
\arguments{
\item{n}{The number of obervations to sample}

\item{sigma}{The covariance matrix}

\item{mean_vector}{The mean vector}
}
\value{
n observations of the p-multivariate X_p ~ N_p(mean_vector, sigma) stored in a n x p matrix
}
\description{
First we compute the cholesky factorization of the covariance
matrix and then we take an affine extension of
A U ~ N(0, 1)
}
\examples{
rmvnorm(10000, matrix(c(1, 0.5, 0.5, 1), 2, 2), rep(0, 2))
}
