% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{rmvnorm_chol}
\alias{rmvnorm_chol}
\title{Simulate a gaussian vector from a lower triangular matrix
that is the result of a cholesky factorization of Sigma}
\usage{
rmvnorm_chol(n, L, mean_vector)
}
\description{
Simulate a gaussian vector from a lower triangular matrix
that is the result of a cholesky factorization of Sigma
}
