% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{rposdef}
\alias{rposdef}
\title{Generate a symmetric positive definite matrix by multiplying a matrix sampled from the goe with its transpose.}
\usage{
rposdef(m)
}
\description{
Any invertible matrix multiplied by its transpose is positive definite. Therefore, we sample a matrix from the goe
and then we multiply it with its transpose to get a matrix that is positive definite. Since singular matrices generated
randomly are extremely rare, there is a high probability that the function returned is indeed positive definite.
}
