% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.r
\name{symmetrize}
\alias{symmetrize}
\title{Make a matrix symmetric}
\usage{
symmetrize(H)
}
\arguments{
\item{H}{The matrix to symmetrize}
}
\value{
The symmetric matrix (H + H')/2
}
\description{
This routine is used to create a matrix variate that is a part of the Gaussian Orthogonal Ensemble
}
\examples{
symmetrize(matrix(c(1, 2, 3, 4), 2, 2)) => matrix(c(1, 2.5, 2.5, 4))
}
